error fix

+++ src/materialDetails.java
@@ -182,17 +182,21 @@ public class materialDetails extends javax.swing.JFrame {
     }//GEN-LAST:event_formWindowActivated
 
     private void jButton3ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton3ActionPerformed
-        int row = jTable1.getSelectedRow();
-        String id = jTable1.getModel().getValueAt(row, 0).toString();
 
-        try {
-            Statement s = DBConnect.getConnection().createStatement();
-            s.executeUpdate("DELETE FROM material where id = '" + id + "'");
-            JOptionPane.showMessageDialog(rootPane, "Deleted");
-            ResultSet rs = s.executeQuery("select * from material");
-            jTable1.setModel(DbUtils.resultSetToTableModel(rs));
-        } catch (Exception e) {
-            System.out.println("Exception = " + e);
+        if (jTable1.getSelectionModel().isSelectionEmpty()) {
+            JOptionPane.showMessageDialog(rootPane, "Select first");
+        } else {
+            try {
+                int row = jTable1.getSelectedRow();
+                String id = jTable1.getModel().getValueAt(row, 0).toString();
+                Statement s = DBConnect.getConnection().createStatement();
+                s.executeUpdate("DELETE FROM material where id = '" + id + "'");
+                JOptionPane.showMessageDialog(rootPane, "Deleted");
+                ResultSet rs = s.executeQuery("select * from material");
+                jTable1.setModel(DbUtils.resultSetToTableModel(rs));
+            } catch (Exception e) {
+                System.out.println("Exception = " + e);
+            }
         }
         // TODO add your handling code here:
     }//GEN-LAST:event_jButton3ActionPerformed
